requirements.txt
Django==1.10.6
gunicorn==19.7.1
Pillow==4.0.0
numpy==1.12.1
scipy==0.19.0
matplotlib==2.0.0
celery==4.0.2
python-memcached==1.58
django-modeltranslation==0.12.1
